Typescript-ifying Extensions billing code
