+++ src/gcp/index.js
@@ -1,6 +1,7 @@
 "use strict";
 
 module.exports = {
+  cloudbilling: require("./cloudbilling"),
   cloudfunctions: require("./cloudfunctions"),
   cloudscheduler: require("./cloudscheduler"),
   cloudlogging: require("./cloudlogging"),
